Removing regression functions and refactoring
